+++ fabric/src/main/java/com/illusivesoulworks/veinmining/VeinMiningFabricMod.java
@@ -17,21 +17,20 @@
 
 package com.illusivesoulworks.veinmining;
 
-import com.illusivesoulworks.spectrelib.config.SpectreConfigInitializer;
 import com.illusivesoulworks.veinmining.common.network.CPacketState;
 import com.illusivesoulworks.veinmining.common.veinmining.VeinMiningEvents;
-import com.illusivesoulworks.veinmining.common.veinmining.VeinMiningPlayers;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
 import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 
-public class VeinMiningFabricMod implements ModInitializer, SpectreConfigInitializer {
+public class VeinMiningFabricMod implements ModInitializer {
 
   public static final ResourceLocation STATE_PACKET =
       new ResourceLocation(VeinMiningConstants.MOD_ID, "state");
@@ -43,8 +42,9 @@ public class VeinMiningFabricMod implements ModInitializer, SpectreConfigInitial
     ServerLifecycleEvents.SERVER_STARTED.register(server -> VeinMiningEvents.reloadDatapack());
     ServerLifecycleEvents.END_DATA_PACK_RELOAD.register(
         (server, resourceManager, success) -> VeinMiningEvents.reloadDatapack());
-    ServerTickEvents.END_WORLD_TICK.register(
-        world -> VeinMiningPlayers.validate(world.getGameTime()));
+    ServerTickEvents.END_WORLD_TICK.register(VeinMiningEvents::tick);
+    ServerPlayConnectionEvents.DISCONNECT.register(
+        (handler, server) -> VeinMiningEvents.playerLoggedOut(handler.getPlayer()));
     PlayerBlockBreakEvents.AFTER.register((world, player, pos, state, blockEntity) -> {
       if (player instanceof ServerPlayer serverPlayer) {
         VeinMiningEvents.blockBreak(serverPlayer, pos, state);
@@ -55,10 +55,5 @@ public class VeinMiningFabricMod implements ModInitializer, SpectreConfigInitial
           CPacketState msg = CPacketState.decode(buf);
           server.execute(() -> CPacketState.handle(msg, player));
         });
-  }
-
-  @Override
-  public void onInitializeConfig() {
-    VeinMiningMod.init();
   }
 }