+++ forge/src/main/java/com/illusivesoulworks/veinmining/common/ForgeCommonEventsListener.java
@@ -25,6 +25,7 @@ import net.minecraft.server.packs.resources.SimplePreparableReloadListener;
 import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.event.server.ServerStartedEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
@@ -68,6 +69,14 @@ public class ForgeCommonEventsListener {
 
     if (evt.getPlayer() instanceof ServerPlayer player) {
       VeinMiningEvents.blockBreak(player, evt.getPos(), evt.getState());
+    }
+  }
+
+  @SubscribeEvent
+  public void playerLoggedOut(final PlayerEvent.PlayerLoggedOutEvent evt) {
+
+    if (evt.getEntity() instanceof ServerPlayer player) {
+      VeinMiningEvents.playerLoggedOut(player);
     }
   }
 }