+++ forge/src/main/java/com/illusivesoulworks/veinmining/platform/ForgeClientPlatform.java
@@ -20,6 +20,7 @@ package com.illusivesoulworks.veinmining.platform;
 import com.illusivesoulworks.veinmining.common.network.CPacketState;
 import com.illusivesoulworks.veinmining.common.network.VeinMiningForgeNetwork;
 import com.illusivesoulworks.veinmining.common.platform.services.IClientPlatform;
+import com.illusivesoulworks.veinmining.common.veinmining.VeinMiningKey;
 import com.mojang.blaze3d.platform.InputConstants;
 import net.minecraft.client.KeyMapping;
 
@@ -27,8 +28,8 @@ public class ForgeClientPlatform implements IClientPlatform {
 
   @Override
   public KeyMapping createKeyMapping(int key, String desc, String category) {
-    return new KeyMapping("key.veinmining.activate.desc", InputConstants.UNKNOWN.getValue(),
-        "key.veinmining.category");
+    return new VeinMiningKey.Mapping("key.veinmining.activate.desc",
+        InputConstants.UNKNOWN.getValue(), "key.veinmining.category");
   }
 
   @Override