+++ quilt/src/main/java/com/illusivesoulworks/veinmining/platform/QuiltClientPlatform.java
@@ -20,6 +20,7 @@ package com.illusivesoulworks.veinmining.platform;
 import com.illusivesoulworks.veinmining.VeinMiningQuiltMod;
 import com.illusivesoulworks.veinmining.common.network.CPacketState;
 import com.illusivesoulworks.veinmining.common.platform.services.IClientPlatform;
+import com.illusivesoulworks.veinmining.common.veinmining.VeinMiningKey;
 import com.mojang.blaze3d.platform.InputConstants;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.network.FriendlyByteBuf;
@@ -30,7 +31,7 @@ public class QuiltClientPlatform implements IClientPlatform {
 
   @Override
   public KeyMapping createKeyMapping(int key, String desc, String category) {
-    return new KeyMapping("key.veinmining.activate.desc", InputConstants.UNKNOWN.getValue(),
+    return new VeinMiningKey.Mapping("key.veinmining.activate.desc", InputConstants.UNKNOWN.getValue(),
         "key.veinmining.category");
   }