+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 Prior to version 1.0.0, this project used MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH.
 
+## [1.2.0+1.20.1] - 2023.06.17
+### Added
+- Added `requiredDestroySpeed` to `veinmining-server.toml` to determine the needed speed on a used tool to successfully
+vein mine blocks
+- Added in-game configuration GUI
+### Changed
+- Updated to Minecraft 1.20.1
+- Configuration options have been renamed with new updated comments
+- `Activate Vein Mining` keybinding will instead display `(Disabled by Configuration)` if the current configuration does
+not use the keybinding
+### Removed
+- Removed `maxDistanceBase` and `maxDistancePerLevel` configuration options
+
 ## [1.1.2+1.19.4] - 2023.04.21
 ### Changed
 - Updated to Minecraft 1.19.4